Refactor: Move storage out of the mixin submodule
